Breakdown TestPlan executor into phases
