+++ core/src/main/java/org/wso2/testgrid/core/TestPlanExecutor.java
@@ -21,79 +21,22 @@ package org.wso2.testgrid.core;
 import org.apache.commons.lang.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.wso2.testgrid.automation.TestAutomationException;
-import org.wso2.testgrid.automation.TestEngine;
-import org.wso2.testgrid.automation.exception.ReportGeneratorException;
-import org.wso2.testgrid.automation.exception.ReportGeneratorInitializingException;
-import org.wso2.testgrid.automation.exception.ReportGeneratorNotFoundException;
-import org.wso2.testgrid.automation.exception.ResultParserException;
-import org.wso2.testgrid.automation.executor.TestExecutor;
-import org.wso2.testgrid.automation.executor.TestExecutorFactory;
-import org.wso2.testgrid.automation.parser.ResultParser;
-import org.wso2.testgrid.automation.parser.ResultParserFactory;
-import org.wso2.testgrid.automation.report.ReportGenerator;
-import org.wso2.testgrid.automation.report.ReportGeneratorFactory;
-import org.wso2.testgrid.common.Agent;
-import org.wso2.testgrid.common.Deployer;
-import org.wso2.testgrid.common.DeploymentCreationResult;
-import org.wso2.testgrid.common.GrafanaDashboardHandler;
-import org.wso2.testgrid.common.InfrastructureProvider;
-import org.wso2.testgrid.common.InfrastructureProvisionResult;
 import org.wso2.testgrid.common.Status;
-import org.wso2.testgrid.common.TestCase;
-import org.wso2.testgrid.common.TestGridConstants;
 import org.wso2.testgrid.common.TestPlan;
-import org.wso2.testgrid.common.TestPlanPhase;
 import org.wso2.testgrid.common.TestPlanStatus;
 import org.wso2.testgrid.common.TestScenario;
-import org.wso2.testgrid.common.config.ConfigurationContext;
-import org.wso2.testgrid.common.config.InfrastructureConfig;
-import org.wso2.testgrid.common.config.ScenarioConfig;
-import org.wso2.testgrid.common.config.Script;
-import org.wso2.testgrid.common.exception.DeployerInitializationException;
-import org.wso2.testgrid.common.exception.InfrastructureProviderInitializationException;
-import org.wso2.testgrid.common.exception.TestGridDeployerException;
-import org.wso2.testgrid.common.exception.TestGridInfrastructureException;
-import org.wso2.testgrid.common.exception.UnsupportedDeployerException;
-import org.wso2.testgrid.common.exception.UnsupportedProviderException;
-import org.wso2.testgrid.common.plugins.AWSArtifactReader;
-import org.wso2.testgrid.common.plugins.ArtifactReadable;
-import org.wso2.testgrid.common.plugins.ArtifactReaderException;
-import org.wso2.testgrid.common.util.DataBucketsHelper;
-import org.wso2.testgrid.common.util.FileUtil;
-import org.wso2.testgrid.common.util.S3StorageUtil;
 import org.wso2.testgrid.common.util.StringUtil;
-import org.wso2.testgrid.common.util.TestGridUtil;
-import org.wso2.testgrid.common.util.tinkerer.SyncCommandResponse;
-import org.wso2.testgrid.common.util.tinkerer.TinkererSDK;
 import org.wso2.testgrid.core.exception.TestPlanExecutorException;
+import org.wso2.testgrid.core.phase.DeployPhase;
+import org.wso2.testgrid.core.phase.InfraPhase;
+import org.wso2.testgrid.core.phase.TestPhase;
 import org.wso2.testgrid.dao.TestGridDAOException;
 import org.wso2.testgrid.dao.uow.TestPlanUOW;
 import org.wso2.testgrid.dao.uow.TestScenarioUOW;
-import org.wso2.testgrid.deployment.DeployerFactory;
-import org.wso2.testgrid.infrastructure.InfrastructureProviderFactory;
 
-import java.io.File;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.OutputStream;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Date;
-import java.util.List;
-import java.util.Optional;
-import java.util.Properties;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
-
-import static java.nio.file.StandardOpenOption.APPEND;
-import static java.nio.file.StandardOpenOption.CREATE;
 
 /**
  * This class is responsible for executing the provided TestPlan.
@@ -102,25 +45,30 @@ import static java.nio.file.StandardOpenOption.CREATE;
  */
 public class TestPlanExecutor {
 
-    public static final int LINE_LENGTH = 72;
+    private static final int LINE_LENGTH = 72;
     private static final Logger logger = LoggerFactory.getLogger(TestPlanExecutor.class);
     private static final int MAX_NAME_LENGTH = 52;
-    private static final int CONFIG_CHANGE_SET_RETRY_COUNT = 10;
-    private TestScenarioUOW testScenarioUOW;
-    private TestPlanUOW testPlanUOW;
-    private ScenarioExecutor scenarioExecutor;
+
+    private  InfraPhase infraPhase;
+    private DeployPhase deployPhase;
+    private TestPhase testPhase;
 
     public TestPlanExecutor() {
-        testPlanUOW = new TestPlanUOW();
-        scenarioExecutor = new ScenarioExecutor();
-        testScenarioUOW = new TestScenarioUOW();
+        infraPhase = new InfraPhase();
+        deployPhase = new DeployPhase();
+        testPhase = new TestPhase();
     }
-
     public TestPlanExecutor(ScenarioExecutor scenarioExecutor, TestPlanUOW testPlanUOW,
                             TestScenarioUOW testScenarioUOW) {
-        this.scenarioExecutor = scenarioExecutor;
-        this.testPlanUOW = testPlanUOW;
-        this.testScenarioUOW = testScenarioUOW;
+        infraPhase = new InfraPhase();
+        deployPhase = new DeployPhase();
+        testPhase = new TestPhase();
+        infraPhase.setTestPlanUOW(testPlanUOW);
+        infraPhase.setTestScenarioUOW(testScenarioUOW);
+        deployPhase.setTestPlanUOW(testPlanUOW);
+        deployPhase.setTestScenarioUOW(testScenarioUOW);
+        testPhase.setTestPlanUOW(testPlanUOW);
+        testPhase.setTestScenarioUOW(testScenarioUOW);
     }
 
     /**
@@ -129,882 +77,14 @@ public class TestPlanExecutor {
      * @param testPlan an instance of {@link TestPlan} in which the tests should be executed
      * @throws TestPlanExecutorException thrown when error on executing test plan
      */
-    public boolean execute(TestPlan testPlan, InfrastructureConfig infrastructureConfig)
-            throws TestPlanExecutorException, TestGridDAOException {
+    public boolean execute(TestPlan testPlan) throws TestPlanExecutorException, TestGridDAOException {
         long startTime = System.currentTimeMillis();
 
-        // Provision infrastructure
-        logger.info("");
-        logger.info("----------------------Start of INFRA PHASE--------------------------");
-        logger.info("TestPlan Status: " + testPlan.getStatus().toString());
-        logger.info("TestPlan Phase: " + testPlan.getPhase().toString());
-        logger.info("");
-        testPlan.setPhase(TestPlanPhase.INFRA_PHASE_STARTED);
-        testPlanUOW.persistTestPlan(testPlan);
-        InfrastructureProvisionResult infrastructureProvisionResult = provisionInfrastructure(infrastructureConfig,
-                testPlan);
-        //setup product performance dashboard
-        if (infrastructureProvisionResult.isSuccess()) {
-            persistTestPlanPhase(testPlan, TestPlanPhase.INFRA_PHASE_SUCCEEDED);
-            GrafanaDashboardHandler dashboardSetup = new GrafanaDashboardHandler(testPlan.getId());
-            dashboardSetup.initDashboard();
-        }
-        logger.info("");
-        logger.info("----------------------End of INFRA PHASE--------------------------");
-        logger.info("TestPlan Status: " + testPlan.getStatus().toString());
-        logger.info("TestPlan Phase: " + testPlan.getPhase().toString());
-        logger.info("");
-
-        // Create and set deployment.
-        if (testPlan.getPhase().equals(TestPlanPhase.INFRA_PHASE_SUCCEEDED) &&
-                testPlan.getStatus().equals(TestPlanStatus.RUNNING)) {
-            logger.info("");
-            logger.info("----------------------Start of DEPLOY PHASE--------------------------");
-            logger.info("TestPlan Status: " + testPlan.getStatus().toString());
-            logger.info("TestPlan Phase: " + testPlan.getPhase().toString());
-            logger.info("");
-            persistTestPlanPhase(testPlan, TestPlanPhase.DEPLOY_PHASE_STARTED);
-            DeploymentCreationResult deploymentCreationResult = createDeployment(testPlan,
-                    infrastructureProvisionResult);
-
-            if (!deploymentCreationResult.isSuccess()) {
-                testPlan.setStatus(TestPlanStatus.ERROR);
-                testPlan.setPhase(TestPlanPhase.DEPLOY_PHASE_ERROR);
-                for (ScenarioConfig scenarioConfig : testPlan.getScenarioConfigs()) {
-                    scenarioConfig.setStatus(Status.DID_NOT_RUN);
-                }
-                testPlanUOW.persistTestPlan(testPlan);
-                logger.error(StringUtil.concatStrings(
-                        "Error occurred while performing deployment for test plan", testPlan.getId(),
-                        "Releasing infrastructure..."));
-                releaseInfrastructure(testPlan, infrastructureProvisionResult, deploymentCreationResult);
-
-                printSummary(testPlan, System.currentTimeMillis() - startTime);
-                return false;
-            }
-            persistTestPlanPhase(testPlan, TestPlanPhase.DEPLOY_PHASE_SUCCEEDED);
-            logger.info("");
-            logger.info("----------------------End of DEPLOY PHASE--------------------------");
-            logger.info("TestPlan Status: " + testPlan.getStatus().toString());
-            logger.info("TestPlan Phase: " + testPlan.getPhase().toString());
-            logger.info("");
-            //Append TestPlan id to deployment.properties file
-            Properties tgProperties = new Properties();
-            tgProperties.setProperty("TEST_PLAN_ID", testPlan.getId());
-            persistAdditionalInputs(tgProperties, DataBucketsHelper.getOutputLocation(testPlan)
-                    .resolve(DataBucketsHelper.DEPL_OUT_FILE));
-            // Append inputs from scenarioConfig in testgrid yaml to deployment outputs file
-            Properties sceProperties = new Properties();
-            for (ScenarioConfig scenarioConfig : testPlan.getScenarioConfigs()) {
-                sceProperties.putAll(scenarioConfig.getInputParameters());
-                persistAdditionalInputs(sceProperties, DataBucketsHelper.getOutputLocation(testPlan)
-                        .resolve(DataBucketsHelper.DEPL_OUT_FILE));
-            }
-
-            // Run test scenarios.
-            // Ideally we should have condition here to check if test-plan has DEPLOY_PHASE_SUCCEEDED phase.
-            // But as per the logic, only those test-plans can reach this line. Hence avoiding the condition.
-            logger.info("");
-            logger.info("----------------------Start of TEST PHASE--------------------------");
-            logger.info("TestPlan Status: " + testPlan.getStatus().toString());
-            logger.info("TestPlan Phase: " + testPlan.getPhase().toString());
-            logger.info("");
-            persistTestPlanPhase(testPlan, TestPlanPhase.TEST_PHASE_STARTED);
-            runScenarioTests(testPlan, deploymentCreationResult);
-            if (testPlan.getStatus().equals(TestPlanStatus.RUNNING) &&
-                    testPlan.getPhase().equals(TestPlanPhase.TEST_PHASE_STARTED)) {
-            } else {
-                logger.error("Continuing to PostTestPlanActions bearing erroneous observations for the test-plan "
-                        + testPlan + ". TestPlan Status: " + testPlan.getStatus() + ", TestPlan Phase: "
-                        + testPlan.getPhase());
-            }
-
-            try {
-                //post test plan actions also belongs to the TEST-Phase
-                performPostTestPlanActions(testPlan, deploymentCreationResult);
-            } catch (Throwable e) {
-                //catch throwable here because we need to ensure the test plan life cycle executes fully
-                //even if an error occurs at this stage.
-                logger.error("Unexpected Error occurred while performing post test execution tasks," +
-                        "hence skipping the step and continuing the test plan lifecycle. ", e);
-            }
-            uploadDeploymentOutputsToS3(testPlan);
-            //cleanup
-            releaseInfrastructure(testPlan, infrastructureProvisionResult, deploymentCreationResult);
-        } else {
-            logger.error("INFRA phase was not succeeded for test-plan: " + testPlan.getId() + ". Hence" +
-                    "not starting next phases.");
-            releaseInfrastructure(testPlan, infrastructureProvisionResult);
-        }
-
-        // Test plan completed. Update and persist testplan status
-        updateTestPlanStatusBasedOnResults(testPlan);
-        logger.info("");
-        logger.info("----------------------End of TEST PHASE--------------------------");
-        logger.info("TestPlan Status: " + testPlan.getStatus().toString());
-        logger.info("TestPlan Phase: " + testPlan.getPhase().toString());
-        logger.info("");
+        testPhase.execute(deployPhase.execute(infraPhase.execute(testPlan)));
 
         // Print summary
         printSummary(testPlan, System.currentTimeMillis() - startTime);
         return testPlan.getStatus() == TestPlanStatus.SUCCESS;
-    }
-
-    /**
-     * Upload deployment outputs (thread-dumps, logs, etc.) from the instances to TestGrid S3 storage.
-     * @param testPlan test-plan of the relevant stack
-     */
-    public void uploadDeploymentOutputsToS3(TestPlan testPlan) {
-        String tinkererHost = ConfigurationContext
-                .getProperty(ConfigurationContext.ConfigurationProperties.DEPLOYMENT_TINKERER_REST_BASE_PATH);
-        if (tinkererHost != null && !tinkererHost.isEmpty()) {
-            logger.info("Process to upload deployment-outputs to S3 has started for test-plan " + testPlan.getId());
-            TinkererSDK tinkererSDK = new TinkererSDK();
-            tinkererSDK.setTinkererHost(ConfigurationContext
-                    .getProperty(ConfigurationContext.ConfigurationProperties.DEPLOYMENT_TINKERER_REST_BASE_PATH));
-            //Check if agent configured for the given test-plan.
-            List<String> activeTestPlans = tinkererSDK.getAllTestPlanIds();
-            if (activeTestPlans != null && activeTestPlans.contains(testPlan.getId())) {
-                List<Agent> agentList = tinkererSDK.getAgentListByTestPlanId(testPlan.getId());
-                logger.info("Found " + agentList.size() + " agents for the test-plan.");
-                //Assuming the external deployment is on a Linux operating system.
-                String configureAWSCLI =
-                        "export AWS_ACCESS_KEY_ID=" + ConfigurationContext
-                                .getProperty(ConfigurationContext.ConfigurationProperties.AWS_ACCESS_KEY_ID_TG_BOT)
-                                + "&&" +
-                                "export AWS_SECRET_ACCESS_KEY=" + ConfigurationContext
-                                .getProperty(ConfigurationContext
-                                        .ConfigurationProperties.AWS_ACCESS_KEY_SECRET_TG_BOT) + "&&" +
-                                "export AWS_DEFAULT_REGION=" + ConfigurationContext
-                                .getProperty(ConfigurationContext.ConfigurationProperties.AWS_REGION_NAME) + "&&";
-                String runLogArchiverScript = "sudo sh /usr/lib/log_archiver.sh &&";
-                String s3Location = deriveDeploymentOutputsDirectory(testPlan);
-                if (s3Location != null) {
-                    ExecutorService executorService = Executors.newCachedThreadPool();
-                    agentList.forEach(agent -> {
-                        String uploadLogsToS3 = "aws s3 cp /var/log/product_logs.zip " +
-                                s3Location + "/product_logs_" + agent.getInstanceName() + ".zip &&";
-                        String uploadDumpsToS3 = "aws s3 cp /var/log/product_dumps.zip " +
-                                s3Location + "/product_dumps_" + agent.getInstanceName() + ".zip";
-                        executorService.execute(new TinkererCommand(agent.getAgentId(), testPlan.getId(),
-                                agent.getInstanceName(),
-                                configureAWSCLI + runLogArchiverScript + uploadLogsToS3 + uploadDumpsToS3));
-                    });
-                    executorService.shutdown();
-                    try {
-                        logger.info("Tinkerer commands are sending out to nodes, Wait till the execution is complete." +
-                                " (TIME-OUT: 10 Minutes)");
-                        if (!executorService.awaitTermination(10, TimeUnit.MINUTES)) {
-                            logger.error("Tinkerer commands execution time-out. Gracefully moving to next steps..");
-                            executorService.shutdownNow();
-                        }
-                    } catch (InterruptedException e) {
-                        logger.error("Exception occurred while waiting for tinkerer commands to be completed. " +
-                                "Please note the commands may not have completely executed due to this. Exception: "
-                                + e.getMessage(), e);
-                    }
-
-                    logger.info("S3 path is : " + s3Location);
-                } else {
-                    logger.error("Can not generate S3 location for deployment-outputs of test-plan: " +
-                            testPlan.getId());
-                }
-            } else {
-                logger.error("Cannot download logs from the deployment. No Tinkerer agents got registered with "
-                        + "Tinkerer for this test-plan. Logs will not be available from the Testgrid dashboard. Test "
-                        + "plan id: " + testPlan.getId());
-            }
-        } else {
-            logger.warn("Tinkerer-host is not configured. Hence uploading deployment-outputs to S3 is skipped.");
-        }
-    }
-
-    /**
-     * Derives the deployment outputs directory for a given test-plan
-     * @param testPlan test-plan
-     * @return directory of the
-     */
-    private String deriveDeploymentOutputsDirectory(TestPlan testPlan) {
-        String s3BucketName = ConfigurationContext
-                .getProperty(ConfigurationContext.ConfigurationProperties.AWS_S3_BUCKET_NAME);
-        if (s3BucketName != null && !s3BucketName.isEmpty()) {
-            try {
-                ArtifactReadable artifactReadable = new AWSArtifactReader(ConfigurationContext.
-                        getProperty(ConfigurationContext.ConfigurationProperties.AWS_REGION_NAME), ConfigurationContext.
-                        getProperty(ConfigurationContext.ConfigurationProperties.AWS_S3_BUCKET_NAME));
-                String path = S3StorageUtil.deriveS3DeploymentOutputsDir(testPlan, artifactReadable);
-                path = "s3://" + s3BucketName + "/" + path;
-                return path;
-            } catch (ArtifactReaderException | IOException e) {
-                logger.error("Error occurred while deriving deployment outputs directory for test-plan " +
-                        testPlan, e);
-            }
-        } else {
-            logger.error("S3 bucket name is not configured in the test environment." +
-                    " Hence can not derive deployment-outputs directory.");
-        }
-        return null;
-    }
-
-    /**
-     * Performs the post test plan tasks using the existing deployment and the results.
-     *
-     * @param testPlan                 current test plan
-     * @param deploymentCreationResult results from the current deployment
-     */
-    private void performPostTestPlanActions(TestPlan testPlan, DeploymentCreationResult deploymentCreationResult) {
-        logger.info("");
-        printSeparator(LINE_LENGTH);
-        logger.info("\t\t Performing Post Run Actions");
-        printSeparator(LINE_LENGTH);
-        logger.info("");
-
-        // Compress test outputs to be uploaded to S3
-        final Path outputLocation = DataBucketsHelper.getTestOutputsLocation(testPlan);
-        Path zipFilePath = Paths.get(outputLocation.toString() + TestGridConstants.TESTGRID_COMPRESSED_FILE_EXT);
-        try {
-            Files.deleteIfExists(zipFilePath);
-            FileUtil.compress(outputLocation.toString(), zipFilePath.toString());
-            logger.info("Created the results archive at " + zipFilePath);
-        } catch (IOException e) {
-            //todo add error to db new column
-            persistTestPlanStatus(testPlan, TestPlanStatus.ERROR);
-            logger.error("Error occurred while archiving test results to" + zipFilePath, e);
-        }
-
-        //Note: Being unable to generate report is not considered as a reason to make test-run ERROR since
-        //report generation is deprecated now.
-        logger.info("Generating report for the test plan: " + testPlan.getId());
-        try {
-            ReportGenerator reportGenerator = ReportGeneratorFactory.getReportGenerator(testPlan);
-            reportGenerator.generateReport();
-        } catch (ReportGeneratorNotFoundException e) {
-            logger.info("Report generation skipped..");
-            logger.debug("Could not find a report generator " +
-                    " for TestPlan of " + testPlan.getDeploymentPattern().getProduct().getName() +
-                    ". Test type: " + testPlan.getScenarioConfigs().get(0).getTestType());
-        } catch (ReportGeneratorInitializingException e) {
-            logger.error("Error while initializing the report generators  " +
-                    "for TestPlan of " + testPlan.getDeploymentPattern().getProduct().getName(), e);
-        } catch (ReportGeneratorException e) {
-            logger.error("Error while generating the report for " +
-                    "TestPlan of " + testPlan.getDeploymentPattern()
-                    .getProduct().getName(), e);
-        }
-    }
-
-    /**
-     * Run all the scenarios mentioned in the testgrid.yaml.
-     *
-     * @param testPlan                 the test plan
-     * @param deploymentCreationResult the result of the previous build step
-     */
-    public void runScenarioTests(TestPlan testPlan, DeploymentCreationResult deploymentCreationResult)
-            throws TestPlanExecutorException {
-
-        for (ScenarioConfig scenarioConfig : testPlan.getScenarioConfigs()) {
-            try {
-                scenarioConfig.setTestPlan(testPlan);
-                TestExecutor testExecutor = TestExecutorFactory.getTestExecutor(
-                        TestEngine.valueOf(scenarioConfig.getTestType()));
-
-                Path scenarioDir = Paths.get(testPlan.getScenarioTestsRepository(), scenarioConfig.getName(),
-                        scenarioConfig.getFile());
-                if (scenarioDir !=  null) {
-                    Path parent = scenarioDir.getParent();
-                    Path file = scenarioDir.getFileName();
-                    if (parent == null) {
-                        parent = Paths.get("");
-                    }
-                    if (file == null) {
-                        file = Paths.get("test.sh");
-                    }
-                    testExecutor.init(parent.toString(), scenarioConfig.getName(), scenarioConfig);
-                    testExecutor.execute(file.toString(), deploymentCreationResult);
-                }
-
-
-            } catch (TestAutomationException e) {
-                //todo: add reason to test-plan db record
-                persistTestPlanProgress(testPlan, TestPlanPhase.TEST_PHASE_INCOMPLETE, TestPlanStatus.ERROR);
-                throw new TestPlanExecutorException("Error while getting test executor for " +
-                        scenarioConfig.getTestType());
-            }
-        }
-        List<TestScenario> testScenarios = new ArrayList<>();
-        testPlan.setTestScenarios(testScenarios);
-        for (ScenarioConfig scenarioConfig : testPlan.getScenarioConfigs()) {
-            populateScenariosList(testPlan, scenarioConfig);
-        }
-
-        for (ScenarioConfig scenarioConfig : testPlan.getScenarioConfigs()) {
-            for (TestScenario testScenario : scenarioConfig.getScenarios()) {
-                populateTestCases(testPlan, testScenario, scenarioConfig);
-                try {
-                    persistTestScenario(testScenario);
-                } catch (TestPlanExecutorException e) {
-                    logger.error(StringUtil.concatStrings(
-                            "Error occurred while persisting test scenario ", testScenario.getName()), e);
-                    //todo: add reason to test-plan db record
-                    persistTestPlanProgress(testPlan, TestPlanPhase.TEST_PHASE_INCOMPLETE, TestPlanStatus.ERROR);
-                }
-            }
-            persistScenarioConfig(scenarioConfig);
-        }
-
-    }
-
-    /**
-     * This method will populate test cases of a give test scenario
-     * @param testPlan          testplan
-     * @param testScenario      scenario of which tests needs to be identified
-     * @param scenarioConfig    scenario config of the test scenario
-     */
-    private void populateTestCases(TestPlan testPlan, TestScenario testScenario, ScenarioConfig scenarioConfig) {
-        Optional<ResultParser> parser = ResultParserFactory.getParser(testPlan, testScenario, scenarioConfig);
-        if (parser.isPresent()) {
-            try {
-                ResultParser resultParser = parser.get();
-                resultParser.parseResults();
-                resultParser.archiveResults();
-            } catch (ResultParserException e) {
-                //todo: add reason to test-plan db record
-                persistTestPlanProgress(testPlan, TestPlanPhase.TEST_PHASE_INCOMPLETE, TestPlanStatus.ERROR);
-                logger.error("Error parsing the results for the scenario " + testScenario.getName(), e);
-            }
-        } else {
-            //todo: add reason to test-plan db record
-            persistTestPlanProgress(testPlan, TestPlanPhase.TEST_PHASE_INCOMPLETE, TestPlanStatus.ERROR);
-            logger.error("Error parsing the results for the scenario no parser " + testScenario.getName());
-        }
-    }
-
-    /**
-     * This method will find the test scenarios fir a given scenario config
-     * @param testPlan          testplan
-     * @param scenarioConfig    scenarioConfig for which tests need to be identified
-     * @throws TestPlanExecutorException
-     */
-    private void populateScenariosList(TestPlan testPlan, ScenarioConfig scenarioConfig) throws
-            TestPlanExecutorException {
-
-        Path dataBucket = Paths.get(DataBucketsHelper.getOutputLocation(testPlan).toString(),
-                "test-outputs", scenarioConfig.getOutputDir(), TestGridConstants.TEST_RESULTS_SCENARIO_DIR);
-        File[] directories = new File(dataBucket.toString()).listFiles(File::isDirectory);
-
-        if (directories != null) {
-            for (File scenario : directories) {
-                if (scenario.getName().equals(TestGridConstants.MAVEN_RELATED_DIR)) {
-                    logger.warn("Avoid considering a directory with maven-reserved name (" +
-                            scenario.getName() + ") as a scenario directory.");
-                } else {
-                    appendScenario(testPlan, scenario.getName(), scenarioConfig);
-                }
-            }
-        } else {
-            //todo: add reason to test-plan db record
-            persistTestPlanProgress(testPlan, TestPlanPhase.TEST_PHASE_INCOMPLETE, TestPlanStatus.ERROR);
-            logger.error("No scenarios found in " + dataBucket + " for Scenario Config " + scenarioConfig.getName() +
-                    " in testplan " + testPlan.getId());
-        }
-
-        logger.info("--------- Identified Scenarios ---------------------");
-        for (TestScenario scenario : scenarioConfig.getScenarios()) {
-            logger.info(scenario.getName());
-        }
-        logger.info("-------------------------------------------------");
-    }
-
-    /**
-     * Append a test scenario to the testplan
-     * @param testPlan          testplan
-     * @param scenarioName      name of the new scenario
-     * @param scenarioConfig    scenario config which is associated with scenario
-     */
-    private void appendScenario(TestPlan testPlan, String scenarioName, ScenarioConfig scenarioConfig) {
-        List<TestScenario> testScenarios = testPlan.getTestScenarios();
-        List<TestScenario> testScenariosOfConfig = scenarioConfig.getScenarios();
-        TestScenario newScenario = new TestScenario();
-        newScenario.setStatus(Status.RUNNING);
-        newScenario.setName(scenarioName);
-        newScenario.setTestPlan(testPlan);
-        newScenario.setIsPostScriptSuccessful(false);
-        newScenario.setIsPreScriptSuccessful(false);
-        newScenario.setConfigChangeSetName("default");
-        newScenario.setDescription(scenarioName);
-        newScenario.setConfigChangeSetDescription(scenarioName);
-        newScenario.setDir(scenarioConfig.getDir());
-        newScenario.setOutputDir(scenarioConfig.getOutputDir());
-        testScenarios.add(newScenario);
-        testScenariosOfConfig.add(newScenario);
-        testPlan.setTestScenarios(testScenarios);
-        scenarioConfig.setScenarios(testScenariosOfConfig);
-
-    }
-
-    /**
-     * Creates the deployment in the provisioned infrastructure.
-     *
-     * @param testPlan                      test plan
-     * @param infrastructureProvisionResult the output of the infrastructure provisioning scripts
-     * @return created {@link DeploymentCreationResult}
-     * @throws TestPlanExecutorException thrown when error on creating deployment
-     */
-    public DeploymentCreationResult createDeployment(TestPlan testPlan,
-            InfrastructureProvisionResult infrastructureProvisionResult)
-            throws TestPlanExecutorException {
-        Path infraOutFilePath = DataBucketsHelper.getOutputLocation(testPlan)
-                .resolve(DataBucketsHelper.INFRA_OUT_FILE);
-        try {
-            logger.info("");
-            printSeparator(LINE_LENGTH);
-            logger.info("\t\t Creating deployment: " + testPlan.getDeploymentConfig().getDeploymentPatterns().get(0)
-                    .getName());
-            printSeparator(LINE_LENGTH);
-            logger.info("");
-
-            if (!infrastructureProvisionResult.isSuccess()) {
-                DeploymentCreationResult result = new DeploymentCreationResult();
-                result.setSuccess(false);
-                logger.debug("Deployment result: " + result);
-                return result;
-            }
-
-            // Append deploymentConfig inputs in testgrid yaml to infra outputs file
-            Properties deplInputs = testPlan.getDeploymentConfig()
-                    .getDeploymentPatterns().get(0).getScripts().get(0).getInputParameters();
-            persistAdditionalInputs(deplInputs, infraOutFilePath);
-
-            Deployer deployerService = DeployerFactory.getDeployerService(testPlan);
-            final DeploymentCreationResult result = deployerService.deploy(testPlan, infrastructureProvisionResult);
-            logger.debug("Deployment result: " + result);
-            return result;
-        } catch (TestGridDeployerException e) {
-            persistTestPlanProgress(testPlan, TestPlanPhase.DEPLOY_PHASE_ERROR, TestPlanStatus.ERROR);
-            String msg = StringUtil
-                    .concatStrings("Exception occurred while running the deployment for deployment pattern '",
-                            testPlan.getDeploymentPattern(), "', in TestPlan");
-            logger.error(msg, e);
-        } catch (DeployerInitializationException e) {
-            persistTestPlanProgress(testPlan, TestPlanPhase.DEPLOY_PHASE_ERROR, TestPlanStatus.ERROR);
-            String msg = StringUtil
-                    .concatStrings("Unable to locate a Deployer Service implementation for deployment pattern '",
-                            testPlan.getDeploymentPattern(), "', in TestPlan '");
-            logger.error(msg, e);
-        } catch (UnsupportedDeployerException e) {
-            persistTestPlanProgress(testPlan, TestPlanPhase.DEPLOY_PHASE_ERROR, TestPlanStatus.ERROR);
-            String msg = StringUtil
-                    .concatStrings("Error occurred while running deployment for deployment pattern '",
-                            testPlan.getDeploymentPattern(), "' in TestPlan");
-            logger.error(msg, e);
-        } catch (Exception e) {
-            // deployment creation should not interrupt other tasks.
-            persistTestPlanProgress(testPlan, TestPlanPhase.DEPLOY_PHASE_ERROR, TestPlanStatus.ERROR);
-            String msg = StringUtil.concatStrings("Unhandled error occurred hile running deployment for deployment "
-                    + "pattern '", testPlan.getDeploymentConfig(), "' in TestPlan");
-            logger.error(msg, e);
-        } finally {
-            // Move infrastructure.properties out of data bucket to avoid exposing to the test execution phase.
-            logger.info("Moving infrastructure.properties from data bucket");
-            File infraOutputFile = infraOutFilePath.toFile();
-            if (!infraOutputFile.renameTo(new File(Paths.get(testPlan.getWorkspace(),
-                    DataBucketsHelper.INFRA_OUT_FILE).toString()))) {
-                logger.error("Error while moving " + infraOutFilePath);
-            }
-        }
-
-        DeploymentCreationResult result = new DeploymentCreationResult();
-        result.setSuccess(false);
-        logger.debug("Deployment result: " + result);
-        return result;
-    }
-
-    /**
-     * Sets up infrastructure for the given {@link InfrastructureConfig}.
-     * TODO: Remove infrastructureConfig since it does not need to be a separate param. It's available within testPlan.
-     *
-     * @param infrastructureConfig infrastructure to set up
-     * @param testPlan             test plan
-     */
-    public InfrastructureProvisionResult provisionInfrastructure(InfrastructureConfig infrastructureConfig,
-            TestPlan testPlan) {
-        try {
-            if (infrastructureConfig == null) {
-                persistTestPlanProgress(testPlan, TestPlanPhase.INFRA_PHASE_ERROR, TestPlanStatus.ERROR);
-                throw new TestPlanExecutorException(StringUtil
-                        .concatStrings("Unable to locate infrastructure descriptor for deployment pattern '",
-                                testPlan.getDeploymentPattern(), "', in TestPlan"));
-            }
-            logger.info("");
-            printSeparator(LINE_LENGTH);
-            logger.info("\t\t Provisioning infrastructure: " + infrastructureConfig.getFirstProvisioner().getName());
-            printSeparator(LINE_LENGTH);
-            logger.info("");
-
-            persistInfraInputs(testPlan);
-            InfrastructureProvisionResult provisionResult = new InfrastructureProvisionResult();
-
-            for (Script script : infrastructureConfig.getFirstProvisioner().getScripts()) {
-                if (!Script.Phase.DESTROY.equals(script.getPhase())) {
-                    InfrastructureProvider infrastructureProvider = InfrastructureProviderFactory
-                            .getInfrastructureProvider(script);
-                    infrastructureProvider.init(testPlan);
-                    logger.info("--- executing script: " + script.getName() + ", file: " + script.getFile());
-                    InfrastructureProvisionResult aProvisionResult = infrastructureProvider.provision(testPlan, script);
-                    addTo(provisionResult, aProvisionResult);
-                }
-            }
-
-            provisionResult.setName(infrastructureConfig.getFirstProvisioner().getName());
-            //TODO: remove. deploymentScriptsDir is deprecated now in favor of DeploymentConfig.
-            provisionResult.setDeploymentScriptsDir(Paths.get(testPlan.getDeploymentRepository()).toString());
-            logger.debug("Infrastructure provision result: " + provisionResult);
-            return provisionResult;
-        } catch (TestGridInfrastructureException e) {
-            persistTestPlanProgress(testPlan, TestPlanPhase.INFRA_PHASE_ERROR, TestPlanStatus.ERROR);
-            String msg = StringUtil
-                    .concatStrings("Error on infrastructure creation for deployment pattern '",
-                            testPlan.getDeploymentPattern(), "', in TestPlan");
-            logger.error(msg, e);
-        } catch (InfrastructureProviderInitializationException | UnsupportedProviderException e) {
-            persistTestPlanProgress(testPlan, TestPlanPhase.INFRA_PHASE_ERROR, TestPlanStatus.ERROR);
-            String msg = StringUtil
-                    .concatStrings("No Infrastructure Provider implementation for deployment pattern '",
-                            testPlan.getDeploymentPattern(), "', in TestPlan");
-            logger.error(msg, e);
-        } catch (RuntimeException e) {
-            // Catching the Exception here since we need to catch and gracefully handle all exceptions.
-            persistTestPlanProgress(testPlan, TestPlanPhase.INFRA_PHASE_ERROR, TestPlanStatus.ERROR);
-            logger.error("Runtime exception while provisioning the infrastructure: " + e.getMessage(), e);
-        } catch (Exception e) {
-            // Catching the Exception here since we need to catch and gracefully handle all exceptions.
-            persistTestPlanProgress(testPlan, TestPlanPhase.INFRA_PHASE_ERROR, TestPlanStatus.ERROR);
-            logger.error("Unknown exception while provisioning the infrastructure: " + e.getMessage(), e);
-        } finally {
-            // Move testplan-props.properties out of data bucket to avoid exposing to the test execution phase.
-            Path testplanPropsFilePath = DataBucketsHelper.getInputLocation(testPlan)
-                    .resolve(DataBucketsHelper.TESTPLAN_PROPERTIES_FILE);
-            logger.info("Moving testplan.properties file from data bucket");
-            File testPlanPropsFile = testplanPropsFilePath.toFile();
-            if (!testPlanPropsFile.renameTo(new File(Paths.get(testPlan.getWorkspace(),
-                    DataBucketsHelper.TESTPLAN_PROPERTIES_FILE).toString()))) {
-                logger.error("Error while moving " + testPlanPropsFile);
-            }
-        }
-
-        InfrastructureProvisionResult infrastructureProvisionResult = new InfrastructureProvisionResult();
-        infrastructureProvisionResult.setSuccess(false);
-        logger.debug("Infrastructure provision result: " + infrastructureProvisionResult);
-        return infrastructureProvisionResult;
-    }
-
-    private void addTo(InfrastructureProvisionResult provisionResult, InfrastructureProvisionResult aProvisionResult) {
-        provisionResult.getProperties().putAll(aProvisionResult.getProperties());
-        if (!aProvisionResult.isSuccess()) {
-            provisionResult.setSuccess(false);
-        }
-    }
-
-    /**
-     * The infra-provision.sh / deploy.sh / run-scenario.sh receive the test plan
-     * configuration as a properties file.
-     *
-     * @param testPlan test plan
-     */
-    private void persistInfraInputs(TestPlan testPlan) {
-        final Path location = DataBucketsHelper.getInputLocation(testPlan)
-                .resolve(DataBucketsHelper.TESTPLAN_PROPERTIES_FILE);
-        final Properties jobProperties = testPlan.getJobProperties();
-        final String keyFileLocation = testPlan.getKeyFileLocation();
-        final Properties infraParameters = testPlan.getInfrastructureConfig().getParameters();
-        try (OutputStream os = Files.newOutputStream(location, CREATE, APPEND)) {
-            jobProperties.store(os, null);
-            infraParameters.store(os, null);
-            os.write((TestGridConstants.KEY_FILE_LOCATION + "=" + keyFileLocation).getBytes(StandardCharsets.UTF_8));
-        } catch (IOException e) {
-            logger.error("Error while persisting infra input params to " + location, e);
-        }
-    }
-
-    /**
-     * Persist additional inputs required other than the outputs from previous steps (i.e. infra/deployment).
-     * The additional inputs are specified in the testgrid.yaml.
-     *
-     * @param properties properties to be added
-     * @param propFilePath path of the property file
-     * @throws TestPlanExecutorException if writing to the property file fails
-     */
-    private void persistAdditionalInputs(Properties properties, Path propFilePath) throws TestPlanExecutorException {
-        try (OutputStream outputStream = new FileOutputStream(
-                propFilePath.toString(), true)) {
-            properties.store(outputStream, null);
-        } catch (Throwable e) {
-            throw new TestPlanExecutorException("Error occurred while writing deployment outputs.", e);
-        }
-    }
-
-    /**
-     * Destroys the given {@link InfrastructureConfig}.
-     *
-     * @param testPlan                      test plan
-     * @param infrastructureProvisionResult the infrastructure provisioning result
-     * @param deploymentCreationResult      the deployment creation result
-     * @throws TestPlanExecutorException thrown when error on destroying
-     *                                   infrastructure
-     */
-    public void releaseInfrastructure(TestPlan testPlan,
-            InfrastructureProvisionResult infrastructureProvisionResult,
-            DeploymentCreationResult deploymentCreationResult)
-            throws TestPlanExecutorException {
-        try {
-            InfrastructureConfig infrastructureConfig = testPlan.getInfrastructureConfig();
-            logger.info("");
-            printSeparator(LINE_LENGTH);
-            logger.info("\t\t Releasing infrastructure: " + infrastructureConfig.getFirstProvisioner().getName());
-            printSeparator(LINE_LENGTH);
-            logger.info("");
-
-            if (TestGridUtil.isDebugMode(testPlan)) {
-                printSeparator(LINE_LENGTH);
-                logger.info(TestGridConstants.DEBUG_MODE + " is enabled. NOT RELEASING the infrastructure. The"
-                        + "infrastructure need to be manually released/de-allocated.");
-                printSeparator(LINE_LENGTH);
-                return;
-            }
-            if (!infrastructureProvisionResult.isSuccess() || !deploymentCreationResult.isSuccess()) {
-                logger.error("Execution of previous steps failed. Trying to release the possibly provisioned "
-                        + "infrastructure");
-            }
-
-            for (Script script : infrastructureConfig.getFirstProvisioner().getScripts()) {
-                if (!Script.Phase.CREATE.equals(script.getPhase())) {
-                    InfrastructureProvider infrastructureProvider = InfrastructureProviderFactory
-                            .getInfrastructureProvider(script);
-                    infrastructureProvider.release(infrastructureConfig, testPlan.getInfrastructureRepository(),
-                            testPlan, script);
-                    // Destroy additional infra created for test execution
-                    infrastructureProvider.cleanup(testPlan);
-                }
-            }
-
-
-        } catch (TestGridInfrastructureException e) {
-            throw new TestPlanExecutorException(StringUtil
-                    .concatStrings("Error on infrastructure removal for deployment pattern '",
-                            testPlan.getDeploymentPattern(), "', in TestPlan"), e);
-        } catch (InfrastructureProviderInitializationException | UnsupportedProviderException e) {
-            throw new TestPlanExecutorException(StringUtil
-                    .concatStrings("No Infrastructure Provider implementation for deployment pattern '",
-                            testPlan.getDeploymentPattern(), "', in TestPlan"), e);
-        }
-    }
-
-    /**
-     * Destroys the given {@link InfrastructureConfig}.
-     *
-     * @param testPlan                      test plan
-     * @param infrastructureProvisionResult the infrastructure provisioning result
-     * @throws TestPlanExecutorException thrown when error on destroying
-     *                                   infrastructure
-     */
-    public void releaseInfrastructure(TestPlan testPlan, InfrastructureProvisionResult infrastructureProvisionResult)
-            throws TestPlanExecutorException {
-        try {
-            InfrastructureConfig infrastructureConfig = testPlan.getInfrastructureConfig();
-            logger.info("");
-            printSeparator(LINE_LENGTH);
-            logger.info("\t\t Releasing infrastructure: " + infrastructureConfig.getFirstProvisioner().getName());
-            printSeparator(LINE_LENGTH);
-            logger.info("");
-
-            if (TestGridUtil.isDebugMode(testPlan)) {
-                printSeparator(LINE_LENGTH);
-                logger.info(TestGridConstants.DEBUG_MODE + " is enabled. NOT RELEASING the infrastructure. The"
-                        + "infrastructure need to be manually released/de-allocated.");
-                printSeparator(LINE_LENGTH);
-                return;
-            }
-
-            for (Script script : infrastructureConfig.getFirstProvisioner().getScripts()) {
-                if (!Script.Phase.CREATE.equals(script.getPhase())) {
-                    InfrastructureProvider infrastructureProvider = InfrastructureProviderFactory
-                            .getInfrastructureProvider(script);
-                    infrastructureProvider.release(infrastructureConfig, testPlan.getInfrastructureRepository(),
-                            testPlan, script);
-                    // Destroy additional infra created for test execution
-                    infrastructureProvider.cleanup(testPlan);
-                }
-            }
-
-        } catch (TestGridInfrastructureException e) {
-            throw new TestPlanExecutorException(StringUtil
-                    .concatStrings("Error on infrastructure removal for deployment pattern '",
-                            testPlan.getDeploymentPattern(), "', in TestPlan"), e);
-        } catch (InfrastructureProviderInitializationException | UnsupportedProviderException e) {
-            throw new TestPlanExecutorException(StringUtil
-                    .concatStrings("No Infrastructure Provider implementation for deployment pattern '",
-                            testPlan.getDeploymentPattern(), "', in TestPlan"), e);
-        }
-    }
-
-    /**
-     * Persists the test plan with the status.
-     * Decided by: if at least one test-case or test-scenario or scenario-config has an error,
-     * then the testplan's phase will be TEST_PHASE_ERROR and the status will be ERROR.
-     * @param testPlan TestPlan object to persist
-     */
-    private void updateTestPlanStatusBasedOnResults(TestPlan testPlan) {
-
-        try {
-            boolean isSuccess = true;
-            for (TestScenario testScenario : testPlan.getTestScenarios()) {
-                if (testScenario.getStatus() != Status.SUCCESS) {
-                    isSuccess = false;
-                    if (testScenario.getStatus() == Status.FAIL) {
-                        testPlan.setStatus(TestPlanStatus.FAIL);
-                        testPlan.setPhase(TestPlanPhase.TEST_PHASE_ERROR);
-                        isSuccess = false;
-                        break;
-                    } else if (testScenario.getStatus() == Status.ERROR) {
-                        logger.error("Found erroneous scenario " + testScenario.getName());
-                        testPlan.setStatus(TestPlanStatus.ERROR);
-                        testPlan.setPhase(TestPlanPhase.TEST_PHASE_ERROR);
-                        isSuccess = false;
-                    }
-                }
-            }
-            for (ScenarioConfig scenarioConfig : testPlan.getScenarioConfigs()) {
-                if (scenarioConfig.getStatus() != Status.SUCCESS) {
-                    isSuccess = false;
-                    if (scenarioConfig.getStatus() == Status.FAIL) {
-                        testPlan.setStatus(TestPlanStatus.FAIL);
-                        isSuccess = false;
-                        break;
-                    } else if (scenarioConfig.getStatus() == Status.ERROR) {
-                        logger.error("Found erroneous scenario config: " + scenarioConfig.getName());
-                        testPlan.setStatus(TestPlanStatus.ERROR);
-                        isSuccess = false;
-                    }
-                }
-            }
-            if (isSuccess) {
-                if (testPlan.getStatus().equals(TestPlanStatus.RUNNING) &&
-                        testPlan.getPhase().equals(TestPlanPhase.TEST_PHASE_STARTED)) {
-                   persistTestPlanProgress(testPlan, TestPlanPhase.TEST_PHASE_SUCCEEDED, TestPlanStatus.SUCCESS);
-                } else {
-                    updateFinalTestPlanPhase(testPlan);
-                    logger.info("All the parsed-results are passing. However due to issues in the previous steps," +
-                            "testplan status can not be claimed as success.");
-                }
-            } else {
-                if (testPlan.getStatus().equals(TestPlanStatus.RUNNING) &&
-                        testPlan.getPhase().equals(TestPlanPhase.TEST_PHASE_STARTED)) {
-                    persistTestPlanProgress(testPlan, TestPlanPhase.TEST_PHASE_SUCCEEDED, TestPlanStatus.FAIL);
-                } else {
-                    updateFinalTestPlanPhase(testPlan);
-                    logger.error("One or more of parsed-results can be failing. However due to issues in the" +
-                            " previous steps, testplan status will be claimed as ERROR (instead of FAILED).");
-                }
-            }
-            testPlanUOW.persistTestPlan(testPlan);
-        } catch (TestGridDAOException e) {
-            logger.error("Error occurred while persisting the test plan. ", e);
-        }
-    }
-
-    /**
-     * Persists the test plan with the status.
-     *
-     * @param testPlan TestPlan object to persist
-     * @param status   the status to set
-     */
-    private void persistTestPlanStatus(TestPlan testPlan, TestPlanStatus status) {
-        logger.info("Updating/Persisting testplan status from " + testPlan.getStatus() + " to " + status);
-        try {
-            testPlan.setStatus(status);
-            testPlanUOW.persistTestPlan(testPlan);
-        } catch (TestGridDAOException e) {
-            logger.error("Error occurred while persisting the test plan. ", e);
-        }
-    }
-
-    private void persistTestPlanPhase(TestPlan testPlan, TestPlanPhase phase) {
-        logger.info("Updating/Persisting testplan phasae from " + testPlan.getPhase() + " to " + phase);
-        try {
-            testPlan.setPhase(phase);
-            testPlanUOW.persistTestPlan(testPlan);
-        } catch (TestGridDAOException e) {
-            logger.error("Error occurred while persisting the test plan. ", e);
-        }
-    }
-
-    private void persistTestPlanProgress(TestPlan testPlan, TestPlanPhase phase, TestPlanStatus status) {
-        logger.info("Updating/Persisting testplan status from " + testPlan.getStatus() + " to " + status + " and " +
-                "testplan phrase from " + testPlan.getPhase() + "to " + phase);
-        try {
-            testPlan.setPhase(phase);
-            testPlan.setStatus(status);
-            testPlanUOW.persistTestPlan(testPlan);
-        } catch (TestGridDAOException e) {
-            logger.error("Error occurred while persisting the test plan. ", e);
-        }
-    }
-
-    /**
-     * Persists the scenario config.
-     *
-     * @param scenarioConfig ScenarioConfig object to persist
-     */
-    private void persistScenarioConfig(ScenarioConfig scenarioConfig) {
-        //Persist test scenario
-        if (scenarioConfig.getScenarios().isEmpty()) {
-            scenarioConfig.setStatus(Status.FAIL);
-        } else {
-            for (TestScenario testScenario : scenarioConfig.getScenarios()) {
-                if (Status.FAIL.equals(testScenario.getStatus())) {
-                    scenarioConfig.setStatus(Status.FAIL);
-                    break;
-                } else if (Status.ERROR.equals(testScenario.getStatus())) {
-                    scenarioConfig.setStatus(Status.ERROR);
-                    break;
-                } else {
-                    scenarioConfig.setStatus(Status.SUCCESS);
-                }
-            }
-        }
-
-    }
-
-    /**
-     * Persists the test scenario.
-     *
-     * @param testScenario TestScenario object to persist
-     */
-    private void persistTestScenario(TestScenario testScenario) throws TestPlanExecutorException {
-        //Persist test scenario
-        try {
-            if (testScenario.getTestCases().isEmpty()) {
-                testScenario.setStatus(Status.ERROR);
-            } else {
-                for (TestCase testCase : testScenario.getTestCases()) {
-                    if (Status.FAIL.equals(testCase.getStatus())) {
-                        testScenario.setStatus(Status.FAIL);
-                        break;
-                    } else {
-                        testScenario.setStatus(Status.SUCCESS);
-                    }
-                }
-            }
-
-            testScenarioUOW.persistTestScenario(testScenario);
-            if (logger.isDebugEnabled()) {
-                logger.debug(StringUtil.concatStrings(
-                        "Persisted test scenario ", testScenario.getName(), " with test cases"));
-            }
-        } catch (TestGridDAOException e) {
-            throw new TestPlanExecutorException(StringUtil.concatStrings(
-                    "Error while persisting test scenario ", testScenario.getName(), e));
-        }
     }
 
     /**
@@ -1017,11 +97,25 @@ public class TestPlanExecutor {
     void printSummary(TestPlan testPlan, long totalTime) {
         switch (testPlan.getStatus()) {
             case SUCCESS:
-                logger.info("all tests passed...");
+                logger.info("All tests passed...");
                 break;
             case ERROR:
-                logger.error("There are deployment/test errors...");
-                logger.info("Error summary is coming soon!");
+                switch(testPlan.getPhase()) {
+                    case INFRA_PHASE_ERROR:
+                        logger.info("Error occurred while provisioning infrastructure...");
+                        break;
+                    case DEPLOY_PHASE_ERROR:
+                        logger.info("Error occurred while creating deployment...");
+                        break;
+                    case TEST_PHASE_INCOMPLETE:
+                        logger.info("Skipped several tests/scenarios due to errors...");
+                        break;
+                    case TEST_PHASE_ERROR:
+                        logger.info("Error occurred while executing tests...");
+                        break;
+                    default:
+                            logger.info("Undefined error occurred: (TestPlanPhase: " + testPlan.getPhase() + ").");
+                }
                 break;
             case FAIL:
                 printFailState(testPlan);
@@ -1120,91 +214,6 @@ public class TestPlanExecutor {
         @Override
         public String toString() {
             return this.osCategory;
-        }
-
-    }
-
-    private void updateFinalTestPlanPhase(TestPlan testPlan) {
-        if (testPlan.getStatus().equals(TestPlanStatus.RUNNING)) {
-            //Change the test-plan phase to currentPhase's ERROR stage
-            //Ex:  INFRA_PHASE_STARTED will be updated to INFRA_PHASE_ERROR
-            TestPlanPhase testPlanPhase = testPlan.getPhase();
-            if (testPlanPhase != null) {
-                String testPlanPhaseStr = testPlan.getPhase().toString();
-                String phaseStage = testPlanPhaseStr.substring(testPlanPhaseStr.lastIndexOf("_") + 1);
-                if (!phaseStage.equals("SUCCEEDED")) {
-                    //If the phase is not succeeded, then we assume the error should be in the same phase.
-                    String phaseName = (testPlanPhaseStr.substring(0, testPlanPhaseStr.lastIndexOf('_')));
-                    logger.info("Setting test-plan's phase " + testPlan.getPhase().toString() +
-                            " into: " + phaseName + "_ERROR");
-                    testPlan.setPhase(TestPlanPhase.valueOf(phaseName + "_ERROR"));
-                    testPlan.setStatus(TestPlanStatus.ERROR);
-                } else {
-                    logger.info("Setting next phase's ERROR stage..");
-                    //If the phase is a succeeded one, then we assume the error should have happened in the next phase.
-                    switch(testPlanPhase) {
-                        case PREPARATION_SUCCEEDED:
-                            testPlan.setPhase(TestPlanPhase.INFRA_PHASE_ERROR);
-                            break;
-                        case INFRA_PHASE_SUCCEEDED:
-                            testPlan.setPhase(TestPlanPhase.DEPLOY_PHASE_ERROR);
-                            break;
-                        case DEPLOY_PHASE_SUCCEEDED:
-                            testPlan.setPhase(TestPlanPhase.TEST_PHASE_ERROR);
-                            break;
-                        default:
-                            logger.error("Error in the phase finalizing logic. Please contact TestGrid admin.");
-                            logger.error("Received testplan with the phase " + testPlan.getPhase());
-                    }
-                    testPlan.setStatus(TestPlanStatus.ERROR);
-                }
-            } else {
-                //Assume the phase is null because the job has aborted/failed before setting the first phase.
-                testPlan.setPhase(TestPlanPhase.PREPARATION_ERROR);
-            }
-            logger.info("=============##### UPDATED TestPlan Result ####==========");
-            logger.info("TestPlan:" + testPlan.getId());
-            logger.info("Status:" + testPlan.getStatus());
-            logger.info("Phase:" + testPlan.getPhase());
-        } else {
-            logger.info("=============##### TestPlan Result ####==========");
-            logger.info("TestPlan:" + testPlan.getId());
-            logger.info("Status:" + testPlan.getStatus());
-            logger.info("Phase:" + testPlan.getPhase());
-        }
-        persistTestPlanProgress(testPlan, testPlan.getPhase(), testPlan.getStatus());
-    }
-}
-
-/**
- * Represents an impl of Runnable which will handle tinkerer command to preprare and upload
- * deployment outputs from single tinkerer-agent.
- */
-class TinkererCommand implements  Runnable {
-    private String agentId;
-    private String testPlanId;
-    private String instanceName;
-    private String command;
-    private TinkererSDK tinkererSDK;
-    private Logger logger = LoggerFactory.getLogger(TinkererCommand.class);
-
-    TinkererCommand (String agentId, String testplanId, String instanceName, String command) {
-        this.agentId = agentId;
-        this.testPlanId = testplanId;
-        this.instanceName = instanceName;
-        this.command = command;
-        tinkererSDK = new TinkererSDK();
-        tinkererSDK.setTinkererHost(ConfigurationContext
-                .getProperty(ConfigurationContext.ConfigurationProperties.DEPLOYMENT_TINKERER_REST_BASE_PATH));
-    }
-    @Override
-    public void run() {
-        SyncCommandResponse syncResponse = tinkererSDK.executeCommandSync(agentId, testPlanId, instanceName, command);
-        if (syncResponse.getExitValue() == 0) {
-            logger.info("Successfully executed tinkerer command for instance: " + instanceName);
-        } else {
-            logger.error("Error received for tinkerer command for instance: " + instanceName +
-                    ". Received error response: " + syncResponse.getResponse());
         }
     }
 }
