+++ core/src/main/java/org/wso2/testgrid/core/command/FinalizeRunTestplan.java
@@ -23,7 +23,6 @@ import org.slf4j.LoggerFactory;
 import org.wso2.testgrid.common.Status;
 import org.wso2.testgrid.common.TestGridConstants;
 import org.wso2.testgrid.common.TestPlan;
-import org.wso2.testgrid.common.TestPlanPhase;
 import org.wso2.testgrid.common.TestPlanStatus;
 import org.wso2.testgrid.common.TestScenario;
 import org.wso2.testgrid.common.config.ConfigurationContext;
@@ -151,30 +150,7 @@ public class FinalizeRunTestplan implements Command {
                         break;
                     }
                 }
-                if (testPlan.getStatus().equals(TestPlanStatus.RUNNING)) {
-                    testPlan.setStatus(TestPlanStatus.ERROR);
-                    //Change the test-plan phase to currentPhase's ERROR stage
-                    //Ex:  INFRA_PHASE_STARTED will be updated to INFRA_PHASE_ERROR
-                    TestPlanPhase testPlanPhase = testPlan.getPhase();
-                    if (testPlanPhase != null) {
-                        String testPlanPhaseStr = testPlan.getPhase().toString();
-                        String phaseName = (testPlanPhaseStr.substring(0, testPlanPhaseStr.lastIndexOf('_')));
-                        testPlan.setPhase(TestPlanPhase.valueOf(phaseName + "_ERROR"));
-                        //todo: if failed before beginning of new one (then need to set new phase's ERROR phase).
-                    } else {
-                        //Assume the phase is null because the job has aborted/failed before setting the first phase.
-                        testPlan.setPhase(TestPlanPhase.PREPARATION_ERROR);
-                    }
-                    logger.info("=============##### FINALIZED TestPlan Result ####==========");
-                    logger.info("TestPlan:" + testPlan.getId());
-                    logger.info("Status:" + testPlan.getStatus());
-                    logger.info("Phase:" + testPlan.getPhase());
-                } else {
-                    logger.info("=============##### TestPlan Result ####==========");
-                    logger.info("TestPlan:" + testPlan.getId());
-                    logger.info("Status:" + testPlan.getStatus());
-                    logger.info("Phase:" + testPlan.getPhase());
-                }
+                testPlan = TestGridUtil.updateFinalTestPlanPhase(testPlan);
                 persistTestPlan(testPlan);
             }
             updateProductStatus();
